+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * 【http  】     HttpUtil增加downloadFileFromUrl（pr#1023@Github）
 * 【core  】     增加toEpochMilli方法
 * 【core  】     Validator修改isCitizenId校验（pr#1032@Github）
+* 【core  】     增加PathUtil和FileNameUtil，分离FileUtil中部分方法
 
 ### Bug修复#
 * 【poi   】     修复ExcelBase.isXlsx方法判断问题（issue#I1S502@Gitee）
